refactor(pds): consolidate direct upload strategy
- Move PDS direct upload into the existing direct upload module.

- Reuse shared direct upload helpers for headers, progress, and completion.

- Keep the PDS direct upload tool name and request flow unchanged.
